fnft_nsev: add NFFT evaluation and optimize slow ES scattering

+++ src/private/fnft__nsev_testcases.c
@@ -885,7 +885,8 @@ const REAL error_bounds[6], fnft_nsev_opts_t * const opts) {
         if (K == 0 && K_exact != 0)
             K = K_exact;        
 #ifdef DEBUG
-        printf("K = %ld, K_exact = %ld\n",K,K_exact);
+        printf("K = %lu, K_exact = %lu\n",
+                (unsigned long)K,(unsigned long)K_exact);
 #endif
         bound_states = malloc(K * sizeof(COMPLEX));
         normconsts_and_residues = malloc(2*D * sizeof(COMPLEX));
